Update time variable type and energy calculation
Changed time variable type from double to unsigned int and added static_cast to energy calculation.

+++ Chapter-4/Projects/Project-1:Circuit-Calculator/jouleResultFour.cpp
@@ -14,9 +14,9 @@ void jouleResultFour(int energyOption_four) {
   cin >> resistance;
 
   cout << "Enter the time of the circuit\n";
-  double time {};
+  unsigned int time {};
   cin >> time;
 
-  double energy {current * current * resistance * time};
+  double energy {current * current * resistance * static_cast<double>time};
   cout << "Energy result of the circuit is " << energy << " joules\n";
 }
